Changes code to retain fractional part till end
Edited lines 6 and 7 by changing int to float type and then rounding them off to int.

+++ src/megaavr/Servo.cpp
@@ -3,8 +3,8 @@
 #include <Arduino.h>
 #include <Servo.h>
 
-#define usToTicks(_us)    ((clockCyclesPerMicrosecond() / 16 * _us) / 4)                 // converts microseconds to tick
-#define ticksToUs(_ticks) (((unsigned) _ticks * 16) / (clockCyclesPerMicrosecond() / 4))   // converts from ticks back to microseconds
+#define usToTicks(_us)    (int)((clockCyclesPerMicrosecond() / 16.0 * _us) / 4)                 // converts microseconds to tick
+#define ticksToUs(_ticks) (int)(((unsigned) _ticks * 16.0) / (clockCyclesPerMicrosecond() / 4.0))   // converts from ticks back to microseconds
 
 #define TRIM_DURATION  5                                   // compensation ticks to trim adjust for digitalWrite delays
